LayoutTests/platform/chromium-win/plugins/document-open-expected.txt

CONSOLE MESSAGE: line 1: PLUGIN: DOCUMENT OPEN SUCCESS
